Improved logging
The Twitter API seems to lie about what the since_id does.  It says "Returns results
with an ID greater than", however it seems to return ones that are equal to the given
since.

The logging now prints the items we skip and their timestamps to help make this clear.